[YSQL]  (#559)  Support secondary index in DML
Summary:
Support Postgres secondary index in DML. The index is now updated in-sync with respect to the base table and can be used in queries.
- Currently, only btree index is supported. Index read/write is done through Postgres index access method API.
- When a row is inserted into a table with indexes, the ybctid is returned. All associated indexes are then populated.
- When a row is deleted from the table, the associated index entries are deleted.
- When a row is updated, the original index entries are first deleted. Then index entries for the updated row are inserted.
- To update / delete index entries, the whole row is now read in the subplan under UPDATE / DELETE.

This revision also adds the support to scan a table or index with a partial range key.

Test Plan: Jenkins

Reviewers: neil, mikhail, mihnea

Reviewed By: mihnea

Subscribers: bogdan, bharat

Differential Revision: https://phabricator.dev.yugabyte.com/D6049

+++ src/yb/yql/pggate/pg_tabledesc.h
@@ -50,15 +50,15 @@ class PgTableDesc : public RefCountedThreadSafe<PgTableDesc> {
     return columns_;
   }
 
-  const int32_t num_hash_key_columns() const {
+  const size_t num_hash_key_columns() const {
     return table_->schema().num_hash_key_columns();
   }
 
-  const int32_t num_key_columns() const {
+  const size_t num_key_columns() const {
     return table_->schema().num_key_columns();
   }
 
-  const int32_t num_columns() const {
+  const size_t num_columns() const {
     return table_->schema().num_columns();
   }
